Add simplecov gem for test coverage analysis. Refactor tests, add tests for key sort and source options.
